update to use @rjsf/core
