+++ example/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@react-native-community/slider": "^3.0.0",
+    "@rjsf/core": "^2.2.0",
     "react": "16.11.0",
     "react-native": "0.62.2",
     "rjsf-native": "link:../rjsf-native"
