+++ rjsf-native/src/TextareaWidget.tsx
@@ -1,4 +1,4 @@
-import {WidgetProps} from 'react-jsonschema-form';
+import {WidgetProps} from '@rjsf/core';
 import TextWidget from './TextWidget';
 import React from 'react';
