+++ rjsf-native/src/Theme.tsx
@@ -1,4 +1,4 @@
-import {ThemeProps} from 'react-jsonschema-form';
+import {ThemeProps} from '@rjsf/core';
 import TextWidget from './TextWidget';
 import TextareaWidget from './TextareaWidget';
 import CheckboxWidget from './CheckboxWidget';
@@ -11,25 +11,40 @@ import FieldTemplate from './FieldTemplate';
 import ObjectFieldTemplate from './ObjectFieldTemplate';
 import ArrayFieldTemplate from './ArrayFieldTemplate';
 import ErrorList from './ErrorList';
+import {MockHTMLForm} from './MockHTMLForm';
+
+const widgets = {
+  TextWidget,
+  TextareaWidget,
+  CheckboxWidget,
+  CheckboxesWidget,
+  PasswordWidget,
+  RadioWidget,
+  SelectWidget: RadioWidget,
+  RangeWidget,
+};
+
+const fields = {
+  TitleField,
+};
 
 const Theme: ThemeProps = {
-  widgets: {
-    TextWidget,
-    TextareaWidget,
-    CheckboxWidget,
-    CheckboxesWidget,
-    PasswordWidget,
-    RadioWidget,
-    SelectWidget: RadioWidget,
-    RangeWidget,
-  },
-  fields: {
-    TitleField,
-  },
+  widgets,
+  fields,
   FieldTemplate,
   ObjectFieldTemplate,
   ArrayFieldTemplate,
   ErrorList,
+  tagName: MockHTMLForm,
 };
 
-export default Theme;
+export {
+  Theme,
+  widgets,
+  fields,
+  FieldTemplate,
+  ObjectFieldTemplate,
+  ArrayFieldTemplate,
+  ErrorList,
+  MockHTMLForm,
+};
